py/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.28.0",
    "pyarrow>=21.0.0",
    "pandas>=2.3.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "torchaudio>=2.6.0",
    "num2words>=0.5.14",
    "timm>=1.0.19",
    "accelerate>=1.10.1",
    "transformers>=4.57.3",
    "flash-attn",
    "openai>=1.102.0",
    "python-dotenv>=1.0.1",
    "peft>=0.15.2",
    "pillow>=11.0.0",
    "websockets>=15.0.1",
    "opencv-python>=4.10.0",
    "cbor2>=5.7.1",
    "backoff>=2.2.1",
    "av>=13.2.0",
    "sam3",
    "decord>=0.6.0",
    "pycocotools>=2.0.10",
]

# PyTorch CUDA wheels index
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# Optional CPU index if needed
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]
# Use local transformers clone for SAM3 support (not yet in released version)
transformers = { path = "transformers", editable = true }
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu",    marker = "sys_platform == 'darwin'" },
]
sam3 = { path = "sam3", editable = true }

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "FALSE" }
